kernel 4.9.119

+++ include/linux/cma.h
@@ -31,7 +31,7 @@ typedef struct of_cma_info_s {
 	of_cma_region_t region[MAX_CMA_AREAS];
 } of_cma_info_t;
 
-#endif // defined(CONFIG_RTD129x) && defined(CONFIG_CMA_AREAS)
+#endif // defined(CONFIG_CMA_AREAS)
 
 extern unsigned cma_area_count;
 extern unsigned long totalcma_pages;
@@ -45,6 +45,6 @@ extern int __init cma_declare_contiguous(phys_addr_t base,
 extern int cma_init_reserved_mem(phys_addr_t base, phys_addr_t size,
 					unsigned int order_per_bit,
 					struct cma **res_cma);
-extern struct page *cma_alloc(struct cma *cma, unsigned int count, unsigned int align);
+extern struct page *cma_alloc(struct cma *cma, size_t count, unsigned int align);
 extern bool cma_release(struct cma *cma, const struct page *pages, unsigned int count);
 #endif
